agents/scoring.py
"""
Scoring Agent (A4)

Two jobs:
1. FAISS similarity search → find K most similar recently-sold properties
2. Livability score → composite neighborhood rating (transit, safety, etc.)

The FAISS part is the technical highlight — each property is encoded into
a 48-dimensional vector, and we find the 10 nearest neighbors in <1ms.
In demo mode, we skip FAISS and use a hardcoded set of comparables.

The livability weights were the hardest thing to calibrate. I surveyed
~40 people (friends, Reddit r/Netherlands, colleagues) about what matters
most when choosing a neighborhood. Transit won by a mile. Schools only
matter if you have kids, so it's weighted lower by default but users
can override via their preference profile.
"""

import math
from typing import Optional

import numpy as np

from agents.base import BaseAgent
from agents.schemas import (
    Comparable,
    FeatureVector,
    LivabilityBreakdown,
    PipelineState,
    PricePrediction,
    ScoredProperty,
)


# ── Livability weights ────────────────────────────────────────────────
# Source: informal survey of ~40 people (biased toward urban young professionals)

DEFAULT_LIVABILITY_WEIGHTS = {
    "transit": 0.30,    # overwhelming #1 priority
    "safety": 0.22,     # CBS crime stats per PC4
    "amenities": 0.20,  # supermarkets, restaurants within walking distance
    "green": 0.18,      # parks, water, green space
    "schools": 0.10,    # lower default — user-adjustable for families
}


def _score_transit(dist_to_station_km: float) -> float:
    """Score transit access based on distance to nearest station.

    < 0.3km → 10/10 (you can hear the train)
    0.3-0.5km → 9/10
    0.5-1.0km → 7-8/10 (still walkable)
    1.0-2.0km → 4-6/10 (bikeable)
    > 2.0km → 2-3/10 (car-dependent)
    """
    if dist_to_station_km <= 0.3:
        return 10.0
    elif dist_to_station_km <= 0.5:
        return 9.0
    elif dist_to_station_km <= 1.0:
        return round(9.0 - (dist_to_station_km - 0.5) * 4, 1)
    elif dist_to_station_km <= 2.0:
        return round(7.0 - (dist_to_station_km - 1.0) * 3, 1)
    else:
        return max(1.0, round(4.0 - (dist_to_station_km - 2.0), 1))


def _score_green(green_pct: float) -> float:
    """Score green space based on % within 500m radius.

    Amsterdam average is ~12%. Vondelpark area is ~25%.
    """
    return round(min(10.0, green_pct * 0.5), 1)


def _score_amenities(supermarket_count: int, restaurant_count: int = 0) -> float:
    """Score amenity access."""
    # Supermarkets matter more (daily need) than restaurants
    super_score = min(5.0, supermarket_count * 1.5)
    resto_score = min(5.0, restaurant_count * 0.2)
    return round(super_score + resto_score, 1)


def _score_safety(pc4: str) -> float:
    """Score safety based on CBS crime statistics.

    In production, this queries CBS Statline API for registered
    crime per PC4. For demo, we use hardcoded scores.
    """
    # CBS-inspired scores (higher = safer)
    safety_scores = {
        "1072": 7.1,  # De Pijp — moderate
        "1015": 6.5,  # Jordaan — tourist area, some petty crime
        "1018": 7.4,  # Oost — relatively quiet
        "1073": 7.2,  # De Pijp south
        "3511": 6.8,  # Utrecht centrum
    }
    return safety_scores.get(pc4, 7.0)


def _score_schools(dist_to_school_km: float) -> float:
    """Score school proximity."""
    if dist_to_school_km <= 0.3:
        return 10.0
    elif dist_to_school_km <= 0.5:
        return 8.0
    elif dist_to_school_km <= 1.0:
        return round(8.0 - (dist_to_school_km - 0.5) * 4, 1)
    else:
        return max(2.0, round(6.0 - dist_to_school_km, 1))


def compute_livability(
    features: dict,
    weights: Optional[dict] = None,
) -> LivabilityBreakdown:
    """Compute composite livability score from features."""
    w = weights or DEFAULT_LIVABILITY_WEIGHTS

    transit = _score_transit(features.get("dist_to_nearest_station_km", 2.0))
    green = _score_green(features.get("green_space_pct_500m", 10.0))
    amenities = _score_amenities(
        features.get("supermarket_count_500m", 2),
        features.get("restaurant_count_1km", 10),
    )
    safety = _score_safety(str(features.get("postcode_4d", "0000"))[:4])
    schools = _score_schools(features.get("dist_to_nearest_school_km", 1.0))

    composite = round(
        transit * w["transit"]
        + safety * w["safety"]
        + amenities * w["amenities"]
        + green * w["green"]
        + schools * w["schools"],
        1,
    )

    return LivabilityBreakdown(
        transit=transit,
        safety=safety,
        amenities=amenities,
        green=green,
        schools=schools,
        composite=composite,
    )


# ── Demo comparables ─────────────────────────────────────────────────

DEMO_COMPARABLES = {
    "1072": [
        Comparable(listing_id="comp-1", address="Ruysdaelkade 88", sold_price=468000, living_area_m2=74, property_type="apartment", similarity_score=0.12),
        Comparable(listing_id="comp-2", address="Van Ostadestraat 145-II", sold_price=452000, living_area_m2=81, property_type="apartment", similarity_score=0.18),
        Comparable(listing_id="comp-3", address="Ferdinand Bolstraat 22-I", sold_price=495000, living_area_m2=85, property_type="apartment", similarity_score=0.24),
    ],
    "1015": [
        Comparable(listing_id="comp-4", address="Westerstraat 102-H", sold_price=412000, living_area_m2=58, property_type="apartment", similarity_score=0.15),
        Comparable(listing_id="comp-5", address="Elandsgracht 74-II", sold_price=438000, living_area_m2=65, property_type="apartment", similarity_score=0.21),
    ],
    "1018": [
        Comparable(listing_id="comp-6", address="Czaar Peterstraat 82", sold_price=365000, living_area_m2=52, property_type="apartment", similarity_score=0.14),
        Comparable(listing_id="comp-7", address="Eerste Leeghwaterstraat 15", sold_price=382000, living_area_m2=60, property_type="apartment", similarity_score=0.19),
    ],
    "1073": [
        Comparable(listing_id="comp-8", address="Van Woustraat 180-II", sold_price=398000, living_area_m2=70, property_type="apartment", similarity_score=0.13),
        Comparable(listing_id="comp-9", address="Ceintuurbaan 312-I", sold_price=415000, living_area_m2=72, property_type="apartment", similarity_score=0.20),
    ],
    "3511": [
        Comparable(listing_id="comp-10", address="Oudegracht 120", sold_price=445000, living_area_m2=88, property_type="apartment", similarity_score=0.22),
        Comparable(listing_id="comp-11", address="Twijnstraat 48", sold_price=390000, living_area_m2=75, property_type="apartment", similarity_score=0.17),
    ],
}


class ScoringAgent(BaseAgent):
    """Agent A4: Score properties with livability + comparables.

    In production, uses FAISS for similarity search. In demo mode,
    uses hardcoded comparables from DEMO_COMPARABLES above.
    """

    def __init__(self):
        super().__init__()
        self._faiss_index = None
        self._load_faiss_index()

    @property
    def name(self) -> str:
        return "scoring"

    def _load_faiss_index(self):
        """Try to load FAISS index. Fail silently for demo mode."""
        try:
            import faiss
            from pathlib import Path

            index_path = Path("models/artifacts/property_index.faiss")
            if index_path.exists():
                self._faiss_index = faiss.read_index(str(index_path))
                self.logger.info("faiss_index_loaded", vectors=self._faiss_index.ntotal)
        except (ImportError, Exception) as e:
            self.logger.info("faiss_not_available_using_demo", reason=str(e)[:100])

    def _get_comparables(self, features: dict, listing_id: str) -> list[Comparable]:
        """Find comparable properties via FAISS or demo fallback."""
        if self._faiss_index is not None:
            # Real FAISS query would go here
            # vec = self._encode_features(features)
            # distances, indices = self._faiss_index.search(vec.reshape(1, -1), k=10)
            # return self._hydrate_comparables(indices[0], distances[0])
            pass

        # Demo mode: return hardcoded comparables for the PC4 area
        pc4 = str(features.get("asking_price", 0))  # hack: we stored postcode in features
        # Try to infer PC4 from lat/lng proximity to known areas
        lat = features.get("lat", 52.37)
        lng = features.get("lng", 4.90)

        # Simple: pick the closest known PC4 area
        best_pc4 = "1072"  # default
        best_dist = float("inf")
        pc4_centers = {
            "1072": (52.353, 4.887),
            "1015": (52.379, 4.882),
            "1018": (52.369, 4.924),
            "1073": (52.350, 4.901),
            "3511": (52.089, 5.118),
        }
        for pc, (plat, plng) in pc4_centers.items():
            d = math.sqrt((lat - plat) ** 2 + (lng - plng) ** 2)
            if d < best_dist:
                best_dist = d
                best_pc4 = pc

        return DEMO_COMPARABLES.get(best_pc4, [])[:8]

    async def _execute(self, state: PipelineState) -> PipelineState:
        if not state.predictions:
            self.logger.info("no_predictions_to_score")
            return state

        fv_store: dict = getattr(state, "_feature_vectors", {})
        raw_store: dict = getattr(state, "_raw_listings", {})
        scored: list[ScoredProperty] = []

        for pred in state.predictions:
            fv: Optional[FeatureVector] = fv_store.get(pred.listing_id)
            if fv is None:
                continue

            features = fv.features

            # Get postcode for safety scoring
            raw_listing = raw_store.get(pred.listing_id)
            pc4 = raw_listing.postcode_4d if raw_listing else "0000"
            features_with_pc = {**features, "postcode_4d": pc4}

            # Livability score
            livability = compute_livability(features_with_pc)

            # Comparable properties
            comparables = self._get_comparables(features, pred.listing_id)

            # Value ratio: predicted / asking (>1.0 means undervalued)
            asking = features.get("asking_price", pred.predicted_price)
            if asking and asking > 0:
                value_ratio = round(pred.predicted_price / asking, 4)
            else:
                value_ratio = 1.0

            scored.append(ScoredProperty(
                listing_id=pred.listing_id,
                asking_price=asking,
                predicted_price=pred.predicted_price,
                ci_lower=pred.ci_lower,
                ci_upper=pred.ci_upper,
                value_ratio=value_ratio,
                livability=livability,
                comparables=comparables,
            ))

        state.scored_properties = scored

        undervalued = sum(1 for s in scored if s.is_undervalued)
        self._log_stats(
            state,
            scored=len(scored),
            undervalued=undervalued,
            avg_livability=round(
                sum(s.livability.composite for s in scored) / max(len(scored), 1), 2
            ),
        )

        return state
