pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "woningscout"
version = "0.9.2"
description = "Autonomous Dutch housing market scout — multi-agent ML system for property valuation and alerts"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Prakhar", email = "prakhar@example.com"},
]

dependencies = [
    "xgboost>=2.0.0",
    "faiss-cpu>=1.7.4",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "langgraph>=0.0.40",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "optuna>=3.5.0",
    "spacy>=3.7.0",
    "geopy>=2.4.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "xxhash>=3.4.0",
    "tenacity>=8.2.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
woningscout-train = "woningscout.train:main"
woningscout-pipeline = "woningscout.pipeline:main"

[tool.setuptools.packages.find]
include = ["agents*", "models*", "services*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
